Fixes - move gdgajubot.gdgajubot to gdgajubot.bot
